+++ AnkiDroid/src/main/java/com/ichi2/anki/AcraCrashReporter.kt
@@ -23,13 +23,15 @@ import androidx.annotation.VisibleForTesting
 import androidx.core.content.edit
 import androidx.core.content.pm.PackageInfoCompat
 import androidx.webkit.WebViewCompat
-import com.ichi2.anki.CrashReporter.Companion.FEEDBACK_REPORT_ALWAYS
-import com.ichi2.anki.CrashReporter.Companion.FEEDBACK_REPORT_ASK
-import com.ichi2.anki.CrashReporter.Companion.FEEDBACK_REPORT_KEY
-import com.ichi2.anki.CrashReporter.Companion.FEEDBACK_REPORT_NEVER
 import com.ichi2.anki.analytics.AnkiDroidCrashReportDialog
 import com.ichi2.anki.analytics.UsageAnalytics
 import com.ichi2.anki.analytics.UsageAnalytics.sendAnalyticsException
+import com.ichi2.anki.common.crashreporting.CrashReportService
+import com.ichi2.anki.common.crashreporting.CrashReporter
+import com.ichi2.anki.common.crashreporting.CrashReporter.Companion.FEEDBACK_REPORT_ALWAYS
+import com.ichi2.anki.common.crashreporting.CrashReporter.Companion.FEEDBACK_REPORT_ASK
+import com.ichi2.anki.common.crashreporting.CrashReporter.Companion.FEEDBACK_REPORT_KEY
+import com.ichi2.anki.common.crashreporting.CrashReporter.Companion.FEEDBACK_REPORT_NEVER
 import com.ichi2.anki.common.time.TimeManager
 import com.ichi2.anki.exception.ManuallyReportedException
 import com.ichi2.anki.exception.UserSubmittedException
