+++ AnkiDroid/src/main/java/com/ichi2/anki/DayRolloverHandler.kt
@@ -35,6 +35,7 @@ import anki.collection.OpChanges
 import anki.collection.opChanges
 import com.ichi2.anki.CollectionManager.withOpenColOrNull
 import com.ichi2.anki.android.AnkiBroadcastReceiver
+import com.ichi2.anki.common.crashreporting.CrashReportService
 import com.ichi2.anki.exception.ManuallyReportedException
 import com.ichi2.anki.libanki.EpochSeconds
 import com.ichi2.anki.libanki.sched.Scheduler
