+++ AnkiDroid/src/main/java/com/ichi2/compat/customtabs/CustomTabsFallback.kt
@@ -17,8 +17,8 @@ package com.ichi2.compat.customtabs
 import android.app.Activity
 import android.content.Intent
 import android.net.Uri
-import com.ichi2.anki.CrashReportService
 import com.ichi2.anki.R
+import com.ichi2.anki.common.crashreporting.CrashReportService
 import com.ichi2.anki.showThemedToast
 import com.ichi2.compat.customtabs.CustomTabActivityHelper.CustomTabFallback
