+++ common/android/src/main/java/com/ichi2/anki/common/crashreporting/CrashReportService.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License along with
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.ichi2.anki
+package com.ichi2.anki.common.crashreporting
 
 import android.app.Activity
 import android.content.Context
